Косметика для StringUtil и dllmain.cpp.

+++ Source/Utils/StringUtil.cpp
@@ -1,5 +1,5 @@
 /*
-* (C) 2020-2021 see Authors.txt
+* (C) 2020-2022 see Authors.txt
 *
 * This file is part of MPC-BE.
 *
@@ -78,7 +78,7 @@ void str_replace(std::wstring& s, const std::wstring_view from, const std::wstri
 
 std::string ConvertWideToANSI(const std::wstring& wstr)
 {
-	int count = WideCharToMultiByte(CP_ACP, 0, wstr.c_str(), wstr.length(), nullptr, 0, nullptr, nullptr);
+	int count = WideCharToMultiByte(CP_ACP, 0, wstr.c_str(), (int)wstr.length(), nullptr, 0, nullptr, nullptr);
 	std::string str(count, 0);
 	WideCharToMultiByte(CP_ACP, 0, wstr.c_str(), -1, &str[0], count, nullptr, nullptr);
 	return str;
@@ -86,15 +86,15 @@ std::string ConvertWideToANSI(const std::wstring& wstr)
 
 std::wstring ConvertAnsiToWide(const std::string& str)
 {
-	int count = MultiByteToWideChar(CP_ACP, 0, str.c_str(), str.length(), nullptr, 0);
+	int count = MultiByteToWideChar(CP_ACP, 0, str.c_str(), (int)str.length(), nullptr, 0);
 	std::wstring wstr(count, 0);
-	MultiByteToWideChar(CP_ACP, 0, str.c_str(), str.length(), &wstr[0], count);
+	MultiByteToWideChar(CP_ACP, 0, str.c_str(), (int)str.length(), &wstr[0], count);
 	return wstr;
 }
 
 std::string ConvertWideToUtf8(const std::wstring& wstr)
 {
-	int count = WideCharToMultiByte(CP_UTF8, 0, wstr.c_str(), wstr.length(), nullptr, 0, nullptr, nullptr);
+	int count = WideCharToMultiByte(CP_UTF8, 0, wstr.c_str(), (int)wstr.length(), nullptr, 0, nullptr, nullptr);
 	std::string str(count, 0);
 	WideCharToMultiByte(CP_UTF8, 0, wstr.c_str(), -1, &str[0], count, nullptr, nullptr);
 	return str;
@@ -102,8 +102,8 @@ std::string ConvertWideToUtf8(const std::wstring& wstr)
 
 std::wstring ConvertUtf8ToWide(const std::string& str)
 {
-	int count = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), str.length(), nullptr, 0);
+	int count = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), (int)str.length(), nullptr, 0);
 	std::wstring wstr(count, 0);
-	MultiByteToWideChar(CP_UTF8, 0, str.c_str(), str.length(), &wstr[0], count);
+	MultiByteToWideChar(CP_UTF8, 0, str.c_str(), (int)str.length(), &wstr[0], count);
 	return wstr;
 }
